refactor: separate providers types away from client deps
will allow us to have provider-based utilities that can be imported in
core modules like messages
